# harness/runs/opencode-kimi-seed1/evidence/api-design/openapi.yaml
openapi: 3.1.0
info:
  title: Library Lending API
  version: 1.0.0
  description: API for managing library book lending, returns, reservations, and fines
servers:
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /members/{id}/loans:
    get:
      summary: Get member's active loans
      operationId: getMemberLoans
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of active loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanWithBook'
  /members/{id}/reservations:
    get:
      summary: Get member's reservations
      operationId: getMemberReservations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationWithBook'
...
  /loans:
    post:
      summary: Create a loan (checkout)
      operationId: createLoan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanInput'
      responses:
        '201':
          description: Loan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '402':
          description: Outstanding fines exceed limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Member is suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          $ref: '#/components/responses/Conflict'
  /loans/{id}/return:
    post:
      summary: Return a book
      operationId: returnLoan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Book returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Loan already returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reservations:
    post:
      summary: Create a reservation
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationInput'
      responses:
        '201':
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /reservations/{id}/cancel:
    post:
      summary: Cancel a reservation
      operationId: cancelReservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Reservation cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  schemas:
...
    Loan:
      type: object
      required:
        - id
        - book_id
        - member_id
        - borrowed_at
        - due_at
        - returned_at
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        borrowed_at:
          type: string
          format: date-time
        due_at:
          type: string
          format: date-time
        returned_at:
          type: string
          format: date-time
          nullable: true
    LoanWithBook:
      allOf:
        - $ref: '#/components/schemas/Loan'
        - type: object
          properties:
            book_title:
              type: string
            days_overdue:
              type: integer
              minimum: 0
    CreateLoanInput:
      type: object
      required:
        - member_id
        - book_id
      properties:
        member_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
    Reservation:
      type: object
      required:
        - id
        - book_id
        - member_id
        - queued_at
        - status
        - notified_at
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
        member_id:
          type: string
          format: uuid
        queued_at:
          type: string
          format: date-time
        status:
          type: string
          enum: [waiting, notified, expired, fulfilled, cancelled]
        notified_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    ReservationWithBook:
      allOf:
        - $ref: '#/components/schemas/Reservation'
        - type: object
          properties:
            book_title:
              type: string
            queue_position:
              type: integer
              minimum: 1
    CreateReservationInput:
      type: object
      required:
        - member_id
        - book_id
      properties:
        member_id:
          type: string
          format: uuid
        book_id:
          type: string
          format: uuid
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'